+++ Mirai-CSharp/Models/Messages/VoiceMessage.cs
@@ -7,31 +7,30 @@ namespace Mirai_CSharp.Models
     /// <summary>
     /// 表示语音消息。
     /// </summary>
-    /// <remarks>
-    /// 提前实现。mirai-api-http不支持传递此消息。
-    /// </remarks>
-    [Obsolete("mirai-api-http不支持传递此消息。")]
     public class VoiceMessage : Messages
     {
         public const string MsgType = "Voice";
         /// <summary>
         /// 语音文件名
         /// </summary>
-        [JsonPropertyName("fileName")]
-        public string FileName { get; set; } = null!;
-        /// <summary>
-        /// 语音文件的md5值
-        /// </summary>
         /// <remarks>
-        /// 个人推测
+        /// 此属性不为 <see langword="null"/> 时将忽略 <see cref="Url"/>
         /// </remarks>
-        [JsonPropertyName("md5")]
-        public string Md5 { get; set; } = null!;
+        [JsonPropertyName("voiceId")]
+        public string VoiceId { get; set; } = null!;
         /// <summary>
         /// 用于下载语音的Url
         /// </summary>
+        /// <remarks>
+        /// 此属性不为 <see langword="null"/> 时将忽略 <see cref="Path"/>
+        /// </remarks>
         [JsonPropertyName("url")]
         public string Url { get; set; } = null!;
+        /// <summary>
+        /// 语音文件的路径
+        /// </summary>
+        [JsonPropertyName("path")]
+        public string Path { get; set; } = null!;
         /// <summary>
         /// 初始化 <see cref="VoiceMessage"/> 类的新实例
         /// </summary>
@@ -40,14 +39,14 @@ namespace Mirai_CSharp.Models
         /// <summary>
         /// 初始化 <see cref="VoiceMessage"/> 类的新实例
         /// </summary>
-        /// <param name="fileName">语音文件名</param>
-        /// <param name="md5">语音文件的md5值 (个人推测)</param>
+        /// <param name="voiceId">语音Id</param>
         /// <param name="url">用于下载语音的Url</param>
-        public VoiceMessage(string fileName, string md5, string url) : this()
+        /// <param name="path">语音文件的路径, 相对路径于 plugins/MiraiAPIHTTP/voices</param>
+        public VoiceMessage(string voiceId, string url, string path) : this()
         {
-            FileName = fileName;
-            Md5 = md5;
+            VoiceId = voiceId;
             Url = url;
+            Path = path;
         }
     }
 }
