+++ Mirai-CSharp/Models/UploadTarget.cs
@@ -1,12 +1,9 @@
-﻿using System;
-
-namespace Mirai_CSharp.Models
+﻿namespace Mirai_CSharp.Models
 {
     /// <summary>
-    /// 图片的目标类型
+    /// 上传对象的目标类型
     /// </summary>
-    [Obsolete("请使用 UploadTarget")]
-    public enum PictureTarget
+    public enum UploadTarget
     {
         /// <summary>
         /// 好友
